refac: split functinal tests into multiple files
