+++ functional_tests/base.py
@@ -1,0 +1,13 @@
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+from selenium import webdriver
+
+
+class FunctionalTest(StaticLiveServerTestCase):
+
+    def setUp(self):
+        super().setUp()
+        self.browser = webdriver.Firefox()
+
+    def tearDown(self):
+        self.browser.quit()
+        super().tearDown()
