+++ functional_tests/test_custom.py
@@ -3,38 +3,10 @@ from django.test import LiveServerTestCase
 from selenium.webdriver.common.keys import Keys
 from selenium.common.exceptions import NoSuchElementException
 
-
-class IndexPageTest(LiveServerTestCase):
-    def setUp(self):
-        super().setUp()
-        self.browser = webdriver.Firefox()
-
-    def tearDown(self):
-        self.browser.quit()
-        super().tearDown()
-
-    def test_visit_index_page(self):
-        # enter into index page
-        self.browser.get(self.live_server_url)
-
-        # checkout the 'get started' link
-        link = self.browser.find_element_by_id('id_get_started')
-        self.assertEqual(link.get_attribute(name='href'), self.live_server_url + '/dirs/')
-
-        link.click()
-
-        # go to browse the directories
-        self.assertEqual(self.browser.current_url, self.live_server_url + '/dirs/')
+from .base import FunctionalTest
 
 
-class CustomSiteTestCase(LiveServerTestCase):
-    def setUp(self):
-        super().setUp()
-        self.browser = webdriver.Firefox()
-
-    def tearDown(self):
-        self.browser.quit()
-        super().tearDown()
+class CustomSiteTestCase(FunctionalTest):
 
     def test_create_directories(self):
         # Alice wants to customize the web site, she enters into the custom home page
